# api/core/openapi.yaml
openapi: 3.0.3
info:
  title: Flight Collaboration Platform Core API
  version: 1.0.0
  description: Internal management and Core business boundary. Core owns business truth.
servers:
  - url: /
paths:
...
  /api/v1/task-change-requests:
    get:
      summary: List task change requests in the current management scope
      description: Requests are paginated and are read-only until a manager or admin reviews them.
      operationId: listTaskChangeRequests
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: task_public_id, in: query, schema: {type: string, format: uuid}}
        - {name: status, in: query, schema: {type: string, enum: [pending, approved, rejected, applied, failed]}}
      responses:
        '200':
          description: Paginated task change requests
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskChangeRequestListResponse'}
        '401': {description: Missing or invalid JWT}
        '403': {description: Principal lacks task change management permission or scope}
    post:
      summary: Submit a task change request
      description: A staff member may report an operational conflict, but the task changes only after manager or admin approval.
      operationId: createTaskChangeRequest
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TaskChangeRequestCreate'}
      responses:
        '201':
          description: Change request created in pending state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskChangeRequestResponse'}
        '400': {description: Invalid task change request}
        '401': {description: Missing or invalid JWT}
        '403': {description: Principal is not allowed to submit a change request in the task scope}
        '404': {description: Task or exception not found in the current scope}
  /api/v1/task-change-requests/{publicID}/review:
    patch:
      summary: Approve or reject a task change request
      description: Only a duty manager or system admin can apply a change. Approval and task mutation are one Core transaction.
      operationId: reviewTaskChangeRequest
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/PublicID'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TaskChangeRequestReview'}
      responses:
        '200':
          description: Review result or idempotent replay
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskChangeRequestResponse'}
        '400': {description: Invalid review request}
        '401': {description: Missing or invalid JWT}
        '403': {description: Only manager or admin can review task changes}
        '404': {description: Change request not found}
        '409': {description: Task state or candidate conflict while applying the approved change}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    flightSourceKey:
      type: apiKey
      in: header
      name: X-Flight-Source-Key
  parameters:
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
    PublicID:
      name: publicID
      in: path
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    TaskChangeRequestCreate:
      type: object
      required: [task_public_id, action, reason]
      properties:
        task_public_id: {type: string, format: uuid}
        exception_public_id: {type: string, format: uuid}
        action: {type: string, enum: [pause, reassign, reschedule, cancel, resume]}
        reason: {type: string, minLength: 1, maxLength: 1024}
        target_candidate_public_id: {type: string, format: uuid}
        target_planned_at: {type: string, format: date-time}
        request_id: {type: string, maxLength: 128}
    TaskChangeRequestReview:
      type: object
      required: [decision]
      properties:
        decision: {type: string, enum: [approve, reject]}
        review_note: {type: string, maxLength: 1024}
    TaskChangeRequest:
      type: object
      required: [public_id, task_public_id, action, reason, status, requested_by_public_id, requested_at, request_id, trace_id]
      properties:
        public_id: {type: string, format: uuid}
        task_public_id: {type: string, format: uuid}
        exception_public_id: {type: string, format: uuid}
        action: {type: string, enum: [pause, reassign, reschedule, cancel, resume]}
        reason: {type: string}
        target_candidate_public_id: {type: string, format: uuid}
        target_planned_at: {type: string, format: date-time}
        status: {type: string, enum: [pending, approved, rejected, applied, failed]}
        requested_by_public_id: {type: string, format: uuid}
        requested_at: {type: string, format: date-time}
        reviewed_by_public_id: {type: string, format: uuid}
        reviewed_at: {type: string, format: date-time}
        review_note: {type: string}
        applied_at: {type: string, format: date-time}
        failure_reason: {type: string}
        request_id: {type: string}
        trace_id: {type: string}
    TaskChangeRequestResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data: {$ref: '#/components/schemas/TaskChangeRequest'}
        request_id: {type: string}
        trace_id: {type: string}
    TaskChangeRequestListResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [items, page, page_size, total]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/TaskChangeRequest'}}
            page: {type: integer, minimum: 1}
            page_size: {type: integer, minimum: 1, maximum: 100}
            total: {type: integer, format: int64, minimum: 0}
        request_id: {type: string}
        trace_id: {type: string}
...